refactor: global-wrapper 폴더 변경 (shared -> features)
