+++ src/features/common/components/global-wrapper/index.ts
@@ -1,0 +1,1 @@
+export * from './global-wrapper'
