+++ src/features/common/components/index.ts
@@ -1,1 +1,2 @@
+export * from './global-wrapper'
 export * from './polling-watcher'
