+++ src/shared/config/providers/index.ts
@@ -1,6 +1,5 @@
 export * from './auth-gaurd'
 export * from './chakra-provider'
-export * from './global-wrapper'
 export * from './modal-provider'
 export * from './Providers'
 export * from './tanstack-provider'
